Refactoring: move run_in_task_center out of dump log subcommand
